fix(gr): decode ratingsSum as float64
Goodreads computes ratingsSum server-side and returns fractional values
(e.g. 7651285.5 for legacy id 113436, Eragon) for some popular works.
The generated client declared the field int64, so the whole payload
failed to unmarshal and affected books 500'd on /book/{id}, vanishing
from author payloads and search results entirely.

Readarr-facing resources keep RatingSum int64; the value is truncated
at the single conversion site, mirroring what hardcover.go already does.

Fixes #577, #579, #581 (upstream).

+++ internal/gr.go
@@ -320,7 +320,7 @@ func mapToWorkResource(book gr.BookInfo, work gr.GetBookGetBookByLegacyIdBookWor
 		IsEbook:            book.Details.Format == "Kindle Edition", // TODO: Flush this out.
 		NumPages:           book.Details.NumPages,
 		RatingCount:        book.Stats.RatingsCount,
-		RatingSum:          book.Stats.RatingsSum,
+		RatingSum:          int64(book.Stats.RatingsSum),
 		AverageRating:      book.Stats.AverageRating,
 		URL:                book.WebUrl,
 		// TODO: Omitting release date is a way to essentially force R to hide
